# swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "info": {
        "title": "Concesionarios API.",
        "version": "1.0.0"
    },
    "definitions": {
        "Coche": {
            "type": "object",
            "properties": {
                "modelo": {
                    "type": "string"
                },
                "cv": {
                    "type": "integer"
                },
                "precio": {
                    "type": "integer"
                }
            },
            "required": [
                "modelo",
                "cv",
                "precio"
            ]
        },
        "Concesionario": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string"
                },
                "direccion": {
                    "type": "string"
                },
                "coches": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "$ref": "#/definitions/Coche"
                    }
                }
            },
            "required": [
                "nombre",
                "direccion",
                "coches"
            ]
        }
    },
    "paths": {
        "/concesionarios": {
            "get": {
                "summary": "Get all concesionarios",
                "responses": {
                    "200": {
                        "description": "List of concesionarios.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Concesionario"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create concesionario",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "concesionario",
                        "schema": {
                            "$ref": "#/definitions/Concesionario"
                        },
                        "example": {
                            "nombre": "concesionario1",
                            "direccion": "dirección inventada",
                            "coches": [
                                {
                                    "modelo": "hyundai",
                                    "cv": 2000,
                                    "precio": 34000
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ message: 'ok' }"
                    }
                }
            }
        },
        "/concesionarios/{id}": {
            "get": {
                "summary": "Get a concesionario by id.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get concesionario.",
                        "schema": {
                            "$ref": "#/definitions/Concesionario"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Concesionario",
                "operationId": "updateConcesionario",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "concesionario",
                        "schema": {
                            "$ref": "#/definitions/Concesionario"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ message: 'ok' }"
                    }
                }
            },
            "delete": {
                "summary": "Delete concesionario.",
                "operationId": "deleteConcesionario",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ message: 'ok' }"
                    }
                }
            }
        },
        "/concesionarios/{id}/coches": {
            "get": {
                "summary": "Get all cars de un concesionario",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "lista de coches",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Coche"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create car",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "coche",
                        "schema": {
                            "$ref": "#/definitions/Coche"
                        },
                        "example": {
                            "modelo": "hyundai",
                            "cv": 2000,
                            "precio": 34000
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ message: 'ok' }"
                    }
                }
            }
        },
        "/concesionarios/{id}/coches/{cocheID}": {
            "get": {
                "summary": "Get a car by id.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "cocheID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "car by id",
                        "schema": {
                            "$ref": "#/definitions/Coche"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update one car",
                "operationId": "updateCar",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "cocheID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "body",
                        "name": "coche",
                        "schema": {
                            "$ref": "#/definitions/Coche"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ message: 'ok' }"
                    }
                }
            },
            "delete": {
                "summary": "Delete one car",
                "operationId": "deleteCar",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "cocheID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ message: 'ok' }"
                    }
                }
            }
        }
    }
}